+++ src/app/admin/components/certifiedProduct/vpEditCertifiedProduct.controller.spec.js
@@ -88,5 +88,16 @@
                 expect(vm.isSaving).toBe(true);
             });
         });
-    });
+
+        it('should warn the user if the status will cause developer suspension', function () {
+            expect(vm.willCauseSuspension('Active')).toBe(false);
+            expect(vm.willCauseSuspension('Retired')).toBe(false);
+            expect(vm.willCauseSuspension('Suspended by ONC')).toBe(false);
+            expect(vm.willCauseSuspension('Suspended by ONC-ACB')).toBe(false);
+            expect(vm.willCauseSuspension('Terminated by ONC')).toBe(true);
+            expect(vm.willCauseSuspension('Withdrawn by Developer')).toBe(false);
+            expect(vm.willCauseSuspension('Withdrawn by Developer Under Surveillance/Review')).toBe(true);
+            expect(vm.willCauseSuspension('Withdrawn by ONC-ACB')).toBe(false);
+        });
+    })
 })();