+++ src/app/decertifications/products/products.controller.js
@@ -25,6 +25,10 @@
                 }, function (error) {
                     $log.debug(error);
                 });
+            commonService.getMeaningfulUseUsersAccurateAsOfDate()
+                .then(function (date) {
+                    vm.muuAccurateAsOf = date.data;
+                });
             commonService.getSearchOptions(true)
                 .then(function (result) {
                     vm.acbs = result.certBodyNames;