+++ src/app/decertifications/products/products.controller.spec.js
@@ -7,6 +7,7 @@
             decertifiedProducts: {"recordCount":2,"pageSize":2,"pageNumber":0,"results":[{"id":6571,"testingLabId":null,"testingLabName":null,"chplProductNumber":"CHP-023867","reportFileLocation":"https://www.icsalabs.com/sites/default/files/2014-EHRA314940-2014-0904-00.pdf","sedReportFileLocation":null,"sedIntendedUserDescription":null,"sedTestingEnd":null,"acbCertificationId":"140238R00","classificationType":{"name":"Complete EHR","id":2},"otherAcb":null,"certificationStatus":{"name":"Withdrawn by Developer","id":3},"developer":{"name":"Nemo Capital Partners","id":1036},"product":{"versionId":3981,"name":"1 Connect BuildYourEMR","id":1736,"version":"5.0"},"certificationEdition":{"name":"2014","id":2},"practiceType":{"name":"Ambulatory","id":1},"certifyingBody":{"name":"ICSA Labs","id":6},"certificationDate":1409803200000,"decertificationDate":1409803200000,"ics":null,"sedTesting":null,"qmsTesting":null,"accessibilityCertified":null,"productAdditionalSoftware":"Microsoft HealthVault, Windows Operating System","transparencyAttestation":null,"transparencyAttestationUrl":null,"countCerts":43,"countCqms":14,"countCorrectiveActionPlans":0,"countCurrentCorrectiveActionPlans":0,"countClosedCorrectiveActionPlans":0,"numMeaningfulUse":null},
                                                                                          {"id":6993,"testingLabId":null,"testingLabName":null,"chplProductNumber":"CHP-028100","reportFileLocation":"https://icsalabs.s3.amazonaws.com/TRS/2014-EHRA149577-2015-0908-00.pdf","sedReportFileLocation":null,"sedIntendedUserDescription":null,"sedTestingEnd":0,"acbCertificationId":"150107R00","classificationType":{"name":"Complete EHR","id":2},"otherAcb":null,"certificationStatus":{"name":"Suspended by ONC","id":7},"developer":{"name":"4medica, Inc.","id":4},"product":{"versionId":934,"name":"4medica iEHR Cloud Ambulatory Suite","id":29,"version":"15.10"},"certificationEdition":{"name":"2014","id":2},"practiceType":{"name":"Ambulatory","id":1},"certifyingBody":{"name":"ICSA Labs","id":6},"certificationDate":1441944000000,"decertificationDate":1409803200000,"ics":null,"sedTesting":null,"qmsTesting":null,"accessibilityCertified":null,"productAdditionalSoftware":"Nitor - HISPDirect, ntpd, Lexicomp, Medline Plus, Surescripts, Adobe Acrobat v10","transparencyAttestation":"Affirmative","transparencyAttestationUrl":"http://4medica.com/pdf/4medica_iehr_mu_stage2certified_2016.pdf","countCerts":41,"countCqms":10,"countCorrectiveActionPlans":1,"countCurrentCorrectiveActionPlans":1,"countClosedCorrectiveActionPlans":0,"numMeaningfulUse":50}]},
 
+            muuAccurateAsOfDate: new Date('2017-01-13'),
             modifiedDecertifiedProducts: [
                 {
                     acb: 'ICSA Labs',
@@ -45,6 +46,7 @@
             module('chpl.decertifications', function ($provide) {
                 $provide.decorator('commonService', function ($delegate) {
                     $delegate.getDecertifiedProducts = jasmine.createSpy('getDecertifiedProducts');
+                    $delegate.getMeaningfulUseUsersAccurateAsOfDate = jasmine.createSpy('getMeaningfulUseUsersAccurateAsOfDate');
                     $delegate.getSearchOptions = jasmine.createSpy('getSearchOptions');
                     return $delegate;
                 });
@@ -55,6 +57,7 @@
                 $q = _$q_;
                 commonService = _commonService_;
                 commonService.getDecertifiedProducts.and.returnValue($q.when(mock.decertifiedProducts));
+                commonService.getMeaningfulUseUsersAccurateAsOfDate.and.returnValue($q.when({data: mock.muuAccurateAsOfDate}));
                 commonService.getSearchOptions.and.returnValue($q.when(mock.searchOptions));
 
                 scope = $rootScope.$new();
@@ -97,6 +100,10 @@
 
         it('should generate the smart-table fields', function () {
             expect(vm.modifiedDecertifiedProducts).toEqual(mock.modifiedDecertifiedProducts);
+        });
+
+        it('should know what the muu_accurate_as_of_date is', function () {
+            expect(vm.muuAccurateAsOf).toEqual(mock.muuAccurateAsOfDate);
         });
     });
 })();